utils: Move source files to internal directory
